perf(bgen): compact variant records

+++ crates/genotype/src/bgen/metadata.rs
@@ -1,8 +1,8 @@
 #[derive(Debug)]
 pub(super) struct VariantRecord {
-    pub(super) probability_payload_offset: usize,
-    pub(super) probability_payload_length: usize,
-    pub(super) declared_uncompressed_block_length: usize,
+    pub(super) probability_payload_offset: u64,
+    pub(super) probability_payload_length: u32,
+    pub(super) declared_uncompressed_block_length: u32,
     #[cfg(test)]
     pub(super) resolved_variant_identifier: String,
 }
